src/requirements.txt
aws-cdk-lib==2.223.0   # replace with the latest stable version
constructs==10.4.3     # compatible range for CDK v2
cryptography==46.0.3   # used to generate key pairs
flask==3.1.2           # used to define web services
requests==2.32.5       # used for external requests within web services